# Conflicts:
#	app/Http/Controllers/Auth/RegisteredUserController.php
#	resources/views/frontend/prime_x/withdraw/now.blade.php

+++ database/migrations/2024_08_07_062236_create_rebate_rules_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->unsignedBigInteger('rule_type_id');
-            $table->decimal('rebate_amount',8,2)->default(0.00);
+            $table->double('rebate_amount',8,2)->default(0.00);
             $table->integer('per_lot');
             $table->boolean('status')->default(0);
             $table->timestamps();
